requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
cfgv==3.4.0
colorama==0.4.6
distlib==0.3.8
Django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
filelock==3.13.1
identify==2.5.33
inflection==0.5.1
iniconfig==2.0.0
nodeenv==1.8.0
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
pytest==8.0.0
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
virtualenv==20.25.0
